courses: S4 - Ep 55

+++ libraryapi/src/main/java/udemy/courses/libraryapi/model/Book.java
@@ -3,6 +3,7 @@ package udemy.courses.libraryapi.model;
 import jakarta.persistence.*;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
@@ -30,8 +31,7 @@ public class Book {
     private GenderBook gender;
 
     @Column(precision = 18, scale = 2)
-    private Double price;
-//    private BigDecimal price;
+    private BigDecimal price;
 
     @ManyToOne()
     @JoinColumn(name = "id_author")
